requirements.txt
avro-python3==1.8.2
psycopg2-binary==2.8.6
confluent-kafka==1.5.0
python-decouple==3.3
